Move debugger for testing to the testing part
