--- src/client/debugger/extension/serviceRegistry.ts
@@ -1,22 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-'use strict';
-
-import { IServiceManager } from '../../ioc/types';
-import { LaunchRequestArguments } from '../types';
-import { DebugEnvironmentVariablesHelper, IDebugEnvironmentVariablesService } from './configuration/resolvers/helper';
-import { LaunchConfigurationResolver } from './configuration/resolvers/launch';
-import { IDebugConfigurationResolver } from './configuration/types';
-
-export function registerTypes(serviceManager: IServiceManager): void {
-    serviceManager.addSingleton<IDebugEnvironmentVariablesService>(
-        IDebugEnvironmentVariablesService,
-        DebugEnvironmentVariablesHelper,
-    );
-    serviceManager.addSingleton<IDebugConfigurationResolver<LaunchRequestArguments>>(
-        IDebugConfigurationResolver,
-        LaunchConfigurationResolver,
-        'launch',
-    );
-}
