+++ src/client/testing/common/debugger/debugLauncher.ts
@@ -1,23 +1,23 @@
 import { inject, injectable, named } from 'inversify';
 import * as path from 'path';
 import { DebugConfiguration, l10n, Uri, WorkspaceFolder } from 'vscode';
-import { IApplicationShell, IDebugService } from '../../common/application/types';
-import { EXTENSION_ROOT_DIR } from '../../common/constants';
-import * as internalScripts from '../../common/process/internal/scripts';
-import { IConfigurationService, IPythonSettings } from '../../common/types';
-import { DebuggerTypeName } from '../../debugger/constants';
-import { IDebugConfigurationResolver } from '../../debugger/extension/configuration/types';
-import { DebugPurpose, LaunchRequestArguments } from '../../debugger/types';
-import { IServiceContainer } from '../../ioc/types';
-import { traceError } from '../../logging';
-import { TestProvider } from '../types';
-import { ITestDebugLauncher, LaunchOptions } from './types';
-import { getConfigurationsForWorkspace } from '../../debugger/extension/configuration/launch.json/launchJsonReader';
-import { getWorkspaceFolder, getWorkspaceFolders } from '../../common/vscodeApis/workspaceApis';
-import { showErrorMessage } from '../../common/vscodeApis/windowApis';
-import { createDeferred } from '../../common/utils/async';
-import { pythonTestAdapterRewriteEnabled } from '../testController/common/utils';
-import { addPathToPythonpath } from './helpers';
+import { IApplicationShell, IDebugService } from '../../../common/application/types';
+import { EXTENSION_ROOT_DIR } from '../../../common/constants';
+import * as internalScripts from '../../../common/process/internal/scripts';
+import { IConfigurationService, IPythonSettings } from '../../../common/types';
+import { DebuggerTypeName } from './constants';
+import { IDebugConfigurationResolver } from './resolvers/types';
+import { DebugPurpose, LaunchRequestArguments } from '../../../debugger/types';
+import { IServiceContainer } from '../../../ioc/types';
+import { traceError } from '../../../logging';
+import { TestProvider } from '../../types';
+import { ITestDebugLauncher, LaunchOptions } from '../types';
+import { getConfigurationsForWorkspace } from './launchJsonReader';
+import { getWorkspaceFolder, getWorkspaceFolders } from '../../../common/vscodeApis/workspaceApis';
+import { showErrorMessage } from '../../../common/vscodeApis/windowApis';
+import { createDeferred } from '../../../common/utils/async';
+import { pythonTestAdapterRewriteEnabled } from '../../testController/common/utils';
+import { addPathToPythonpath } from '../helpers';
 
 @injectable()
 export class DebugLauncher implements ITestDebugLauncher {
