+++ src/client/testing/common/debugger/resolvers/base.ts
@@ -16,9 +16,9 @@ import { IInterpreterService } from '../../../../interpreter/contracts';
 import { sendTelemetryEvent } from '../../../../telemetry';
 import { EventName } from '../../../../telemetry/constants';
 import { DebuggerTelemetry } from '../../../../telemetry/types';
-import { DebugOptions, LaunchRequestArguments, PathMapping } from '../../../types';
-import { PythonPathSource } from '../../types';
-import { IDebugConfigurationResolver } from '../types';
+import { DebugOptions, LaunchRequestArguments, PathMapping } from '../../../../debugger/types';
+import { PythonPathSource } from '../../../../debugger/extension/types';
+import { IDebugConfigurationResolver } from './types';
 import { resolveVariables } from '../utils/common';
 import { getProgram } from './helper';
