--- src/test/debugger/common/constants.ts
@@ -1,7 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-'use strict';
-
-// Sometimes PTVSD can take a while for thread & other events to be reported.
-export const DEBUGGER_TIMEOUT = 20000;
